Initial Memory implementation done. Currently Prototyping Cpu
Committing and pushing because I want it backed up mostly. I tryed to
get a mutable Array to work but gave up and figured out how to get an
immutable array to work. It should not matter since performance really
is not a concern.

Eventually there will be some sort of device controller or bridge
between the Cpu and Memory to support other devices (like our web
terminal) but not yet. Not exactly sure how the interface will be with
the Cpu yet and how we are going to run it with a clock. That will have
to be in IO so I will have to figure it out later.

Going to have to hardcode the machine code I have at some point. Not
exactly sure how I am going to do that but we'll see.