Loads of refactoring. This one compiles.
Not sure if this actually works at all. Did a lot of name changing and
moving around of functions to make it more readable. Added a generic
device that all of our hardware will instance from. Will be using a Bus
to send requests to the correct device (based on address). First step is
to refactor the Cpu code to use the Bus instead of hard coded memory.
Going to have to queue up IO requests somehow with the Display device.
Not exactly sure how that is going to work.