Peripherals will not be written in Haskell
This does not make sense in the grand scheme of how this project will
work. They will be on the frontend (web) communicated through web
sockets. They are not part of the state of this machine. The state of
what is in the `Host` will be pure (memory, cpu, hdd, etc). We will
communicate to non-pure peripherals through `Ports`.

So I guess the difference will be a `Device` will be anything that can
get and set bytes in a pure way and a `Port` will be anything that can
get and set bytes in a IO way.

The kernel will get a copy of the pure computer (thus read only) which
will have to be able to take that read only state and send it to the
frontend (web sockets again). Perhaps this won't actually be the kernel
and instead a hard coded step in the main IO loop.